Add files via upload
Scheduled by the interval we give it (the interval rn is 18 minutes.... It won't run at first give it like the time you said the interval was and then it starts and does it again and again and again as long as the job is under the function of miner. The email stuff I havent really finished because apparently I have to login to the server still and wanted to use no login way but didnt work. It will be pretty simple though the email is just to send us the result of the test we will incorporate and literally just get that program started and was completed and here is the result of the test (PASS=same # as website, FAIL=Different number)